Refactored psd plotting for Raw and Epochs
